ifpi_questoes/requirements.txt
asgiref==3.7.2
Django==5.0.3
pillow==10.2.0
sqlparse==0.4.4
django-import-export==3.3.7